fix: [FE] 에디터 컴포넌트 세분화 및 반응형 추가 컴포넌트 쪼개기 및 반응형 추가 (#71)
